Move the status poll endpoint to its own controller
